+++ packages/lib/getFeatures/index.ts
@@ -1,2 +1,6 @@
 export * from './types'
 export { getWfsFeatures } from './wfs'
+export {
+  getVectorFeaturesByBboxRequest,
+  getVectorFeaturesByFeatureRequest,
+} from './vector'